Handle service disabled in configure_init_service
